some changes to get rid of CB warnings... new lines, etc...
git-svn-id: http://svn.openframeworks.cc/openFrameworks/trunk/openFrameworks@267 5aee0c02-c622-0410-bcf2-d8652e674aa7

+++ graphics/ofGraphics.cpp
@@ -437,7 +437,7 @@ void ofDisableSmoothing(){
 //----------------------------------------------------------
 void ofSetStyle(ofStyle style){
 	//color
-	ofSetColor(style.color.r, style.color.g, style.color.b, style.color.a);
+	ofSetColor((int)style.color.r, (int)style.color.g, (int)style.color.b, (int)style.color.a);
 
 	//circle resolution - don't worry it only recalculates the display list if the res has changed
 	ofSetCircleResolution(style.circleResolution);
